refactor([issue-758]): extract subcomponents from StoryboardPanel
Split the 1199-LOC StoryboardPanel.jsx into focused per-tab files so the
panel is a lean orchestrator (375 LOC) over its state/effects and tab
routing:

- StoryboardBibleTab.jsx  — characters/world/objects bible tab + header
- StoryboardScenesTab.jsx — prose-heading outline
- StoryboardBoardsTab.jsx — render dock (scene cards) + status row
- StoryboardBoardsBanners.jsx — setup/failed/stale/missing-bibles banners
- StoryboardConfigTab.jsx — image-gen + world-style + Adapt LLM config

No behavior change; named exports (STORYBOARD_TAB / STORYBOARD_TAB_VALUES)
and the default export are preserved.
